unified-site/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0
langchain==0.3.0
langchain-openai==0.3.0
qdrant-client==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pypdf==4.0.0
ebooklib==0.18
Pillow==10.1.0
openai>=1.3.7
numpy>=1.21.0
psycopg2-binary
cryptography>=41.0.0
httpx==0.25.2
python-slugify==8.0.1
beautifulsoup4==4.12.2
markdown==3.5.1
python-frontmatter==1.1.0
better-exceptions==0.3.2
neon-py==0.0.1